sdk-tests: Move Lago's fixtures to an external module
This way Lago's fixtures can be shared across different
test modules.
